Update README.md
My changes are an update to the process of joining and learning about how to use Github. They are a description of the steps I personally took to get to this point. I am trying to learn how to code. I have now learned to Create a repository with a description and named it hello-world. I have learned to create a Branch from that Master Repository. I have learned to edit the Branch and commit to the changes made. The changes being the edit to the hello-world file.